+++ game/client/c_basecombatcharacter.h
@@ -84,7 +84,7 @@ public:
 	// This is a sort of hack back-door only used by physgun!
 	void SetAmmoCount( int iCount, int iAmmoIndex );
 
-	float				GetNextAttack() const { return m_flNextAttack; }
+	double				GetNextAttack() const { return m_flNextAttack; }
 	void				SetNextAttack( double flWait ) { m_flNextAttack = flWait; }
 
 	virtual int			BloodColor();
